Proj/variable naming pass. Updated automotive.
