+++ custom-commands/automotive/azure-function/AutomotiveApp.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29509.3
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VirtualInventoryApp", "VirtualInventoryApp\VirtualInventoryApp.csproj", "{91B7A47C-93B0-4F63-B478-B68D7A71D266}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AutomotiveApp", "AutomotiveApp\AutomotiveApp.csproj", "{91B7A47C-93B0-4F63-B478-B68D7A71D266}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
