+++ custom-commands/automotive/azure-function/AutomotiveApp/Connections.json
@@ -1,0 +1,3 @@
+{
+  "AZURE_STORAGE_URL": ""
+}
