+++ custom-commands/automotive/azure-function/AutomotiveApp/host.json
