--- custom-commands/automotive/azure-function/VirtualAutomotiveApp/Connections.json
@@ -1,3 +1,0 @@
-{
-  "AZURE_STORAGE_URL": ""
-}
