+++ custom-commands/automotive/visualization/automotive.html
@@ -191,13 +191,6 @@
                     $("#temperature").html(configuration.Temperature + "°");
 
                     if (configuration.Help) {
-                        $("#helpboxes").fadeOut('fast')
-                    }
-                    else {
-                        $("#helpboxes").fadeIn('fast')
-                    }
-
-                    if (configuration.Help) {
                         $("#help-content").fadeIn('fast')
                     }
                     else {
