+++ custom-commands/inventory/azure-function/InventoryApp/Connections.json
@@ -1,0 +1,3 @@
+{
+  "AZURE_STORAGE_URL": ""
+}
