+++ custom-commands/inventory/azure-function/InventoryApp/host.json
