--- custom-commands/inventory/azure-function/VirtualInventoryApp/Connections.json
@@ -1,3 +1,0 @@
-{
-  "AZURE_STORAGE_URL": ""
-}
